requirements.txt
Flask==3.0.0
Pillow==10.2.0
pandas==2.2.0
qrcode==7.4.2
openpyxl==3.1.2
gunicorn==21.2.0